fix(react-redux): Replace deprecated withRef attribute with forwardRef
