+++ client/components/ItemActionConfirmation/forms/createPlanningForm.tsx
@@ -89,5 +89,5 @@ export const CreatePlanningForm = connect(
     mapStateToProps,
     mapDispatchToProps,
     null,
-    {withRef: true}
+    {forwardRef: true}
 )(CreatePlanningComponent);
