+++ client/components/ItemActionConfirmation/forms/editPriorityForm.tsx
@@ -177,5 +177,5 @@ export const EditPriorityForm = connect(
     mapStateToProps,
     mapDispatchToProps,
     null,
-    {withRef: true}
+    {forwardRef: true}
 )(EditPriorityComponent);
