+++ client/components/ItemActionConfirmation/forms/spikePlanningForm.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
+import {get} from 'lodash';
 
 import {appConfig} from 'appConfig';
 
@@ -9,12 +10,13 @@ import '../style.scss';
 import {gettext, getDateTimeString} from '../../../utils';
 import {Row} from '../../UI/Preview';
 
-export class UnspikePlanningComponent extends React.Component {
+export class SpikePlanningComponent extends React.Component {
     constructor(props) {
         super(props);
         // Initialise with enable save so that the user can action on this plan.
         props.enableSaveInModal();
     }
+
 
     submit() {
         return this.props.onSubmit(this.props.original);
@@ -54,18 +56,24 @@ export class UnspikePlanningComponent extends React.Component {
     }
 }
 
-UnspikePlanningComponent.propTypes = {
+SpikePlanningComponent.propTypes = {
     original: PropTypes.object.isRequired,
     onSubmit: PropTypes.func,
     enableSaveInModal: PropTypes.func,
 };
 
 const mapDispatchToProps = (dispatch) => ({
-    onSubmit: (plan) => dispatch(actions.planning.ui.unspike(plan)),
+    onSubmit: (plan) => dispatch(actions.planning.ui.spike(plan)),
+    onHide: (plan, modalProps) => {
+        if (get(modalProps, 'onCloseModal')) {
+            modalProps.onCloseModal(plan);
+        }
+        return Promise.resolve(plan);
+    },
 });
 
-export const UnspikePlanningForm = connect(
+export const SpikePlanningForm = connect(
     null,
     mapDispatchToProps,
     null,
-    {withRef: true})(UnspikePlanningComponent);
+    {forwardRef: true})(SpikePlanningComponent);
