+++ client/components/ItemActionConfirmation/forms/updateRecurringEventsForm.tsx
@@ -2,31 +2,43 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import * as actions from '../../../actions';
+import {get} from 'lodash';
 import {UpdateMethodSelection} from '../UpdateMethodSelection';
 import {EventScheduleSummary, EventUpdateMethods} from '../../Events';
 import {eventUtils, gettext} from '../../../utils';
 import {Row} from '../../UI/Preview';
 import '../style.scss';
 
-export class AssignCalendarComponent extends React.Component {
+export class UpdateRecurringEventsComponent extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             eventUpdateMethod: EventUpdateMethods[0],
             relatedEvents: [],
+            relatedPlannings: [],
         };
 
         this.onEventUpdateMethodChange = this.onEventUpdateMethodChange.bind(this);
     }
 
     componentWillMount() {
-        const event = eventUtils.getRelatedEventsForRecurringEvent(
-            this.props.original,
-            EventUpdateMethods[0],
-            true
-        );
+        const isRecurring = get(this.props, 'original.recurrence_id');
 
-        this.setState({relatedEvents: event._events});
+        if (isRecurring || eventUtils.eventHasPlanning(this.props.original)) {
+            this.posting = get(this.props.original, '_post', true);
+            const event = isRecurring ?
+                eventUtils.getRelatedEventsForRecurringEvent(
+                    this.props.original,
+                    EventUpdateMethods[0],
+                    true
+                ) :
+                this.props.original;
+
+            this.setState({
+                relatedEvents: event._events,
+                relatedPlannings: this.posting ? [] : event._relatedPlannings,
+            });
+        }
 
         // Enable save so that the user can update just this event.
         this.props.enableSaveInModal();
@@ -42,11 +54,11 @@ export class AssignCalendarComponent extends React.Component {
         this.setState({
             eventUpdateMethod: option,
             relatedEvents: event._events,
+            relatedPlannings: this.posting ? [] : event._relatedPlannings,
         });
     }
 
     submit() {
-        // Send the required fields and calendar only
         return this.props.onSubmit(
             this.props.original,
             {
@@ -58,7 +70,11 @@ export class AssignCalendarComponent extends React.Component {
 
     render() {
         const {original, submitting} = this.props;
-        const numEvents = this.state.relatedEvents.length + 1;
+        const isRecurring = !!original.recurrence_id;
+        const eventsInUse = this.state.relatedEvents.filter((e) => (
+            get(e, 'planning_ids.length', 0) > 0 || 'pubstatus' in e
+        ));
+        const numEvents = this.state.relatedEvents.length + 1 - eventsInUse.length;
 
         return (
             <div className="MetadataView">
@@ -77,10 +93,14 @@ export class AssignCalendarComponent extends React.Component {
                     className="strong"
                 />
 
-                <EventScheduleSummary schedule={original.dates} noPadding={true} />
+                <EventScheduleSummary
+                    schedule={original.dates}
+                    forUpdating={true}
+                    useEventTimezone={true}
+                />
 
                 <Row
-                    enabled={true}
+                    enabled={isRecurring}
                     label={gettext('No. of Events')}
                     value={numEvents}
                     noPadding={true}
@@ -89,37 +109,54 @@ export class AssignCalendarComponent extends React.Component {
                 <UpdateMethodSelection
                     value={this.state.eventUpdateMethod}
                     onChange={this.onEventUpdateMethodChange}
-                    showMethodSelection={true}
+                    showMethodSelection={isRecurring}
                     updateMethodLabel={gettext('Update all recurring events or just this one?')}
                     showSpace={false}
                     readOnly={submitting}
+                    action="unpost"
+                    relatedPlannings={this.state.relatedPlannings}
                 />
             </div>
         );
     }
 }
 
-AssignCalendarComponent.propTypes = {
+UpdateRecurringEventsComponent.propTypes = {
     original: PropTypes.object.isRequired,
     updates: PropTypes.object.isRequired,
     submitting: PropTypes.bool,
     onSubmit: PropTypes.func,
     enableSaveInModal: PropTypes.func,
+    resolve: PropTypes.func,
 };
 
-const mapDispatchToProps = (dispatch) => ({
+const mapDispatchToProps = (dispatch, ownProps) => ({
     onSubmit: (original, updates) => (
         dispatch(actions.main.save(original, updates, false))
-            .then((savedItem) => dispatch(actions.events.api.unlock(savedItem)))
+            .then((savedItem) => {
+                if (ownProps.modalProps.unlockOnClose) {
+                    dispatch(actions.events.api.unlock(savedItem));
+                }
+
+                if (ownProps.resolve) {
+                    ownProps.resolve(savedItem);
+                }
+            })
     ),
     onHide: (event) => {
-        dispatch(actions.events.api.unlock(event));
+        if (ownProps.modalProps.unlockOnClose) {
+            dispatch(actions.events.api.unlock(event));
+        }
+
+        if (ownProps.resolve) {
+            ownProps.resolve();
+        }
     },
 });
 
-export const AssignCalendarForm = connect(
+export const UpdateRecurringEventsForm = connect(
     null,
     mapDispatchToProps,
     null,
-    {withRef: true}
-)(AssignCalendarComponent);
+    {forwardRef: true}
+)(UpdateRecurringEventsComponent);
